requirements.txt
yfinance>=0.2.36
pandas>=2.0
numpy>=1.24
matplotlib>=3.7
anthropic>=0.42
pydantic>=2.0
requests>=2.31
streamlit>=1.32
plotly>=5.18
python-dotenv>=1.0
scipy>=1.10
fastapi>=0.110
uvicorn[standard]>=0.29
